backend/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.1
PyJWT>=2.8.0
bcrypt>=4.1.0
opencv-python>=4.5.0
numpy>=1.19.0,<2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
